+++ src/plugins/components/placeholder.css
@@ -1,0 +1,6 @@
+.Placeholder {
+}
+
+.Placeholder-status {
+  text-align: center;
+}
